+++ ikarus/finiteelements/mechanics/linearelastic.hh
@@ -24,6 +24,7 @@
 #  include <ikarus/finiteelements/febases/powerbasisfe.hh>
 #  include <ikarus/finiteelements/fehelper.hh>
 #  include <ikarus/finiteelements/ferequirements.hh>
+#  include <ikarus/finiteelements/mechanics/loads.hh>
 #  include <ikarus/finiteelements/mechanics/materials.hh>
 #  include <ikarus/finiteelements/physicshelper.hh>
 #  include <ikarus/utils/defaultfunctions.hh>
@@ -41,21 +42,26 @@ namespace Ikarus {
    * @tparam useEigenRef A boolean flag indicating whether to use Eigen references.
    */
   template <typename Basis_, typename FERequirements_ = FERequirements<>, bool useEigenRef = false>
-  class LinearElastic : public PowerBasisFE<typename Basis_::FlatBasis> {
+  class LinearElastic : public PowerBasisFE<typename Basis_::FlatBasis>,
+                        public Volume<LinearElastic<Basis_, FERequirements_, useEigenRef>,
+                                      TraitsFromFE<Basis_, FERequirements_, useEigenRef>>,
+                        public Traction<LinearElastic<Basis_, FERequirements_, useEigenRef>,
+                                        TraitsFromFE<Basis_, FERequirements_, useEigenRef>> {
   public:
-    using Basis                   = Basis_;
-    using FlatBasis               = typename Basis::FlatBasis;
-    using BaseDisp                = PowerBasisFE<FlatBasis>;  // Handles globalIndices function
-    using FERequirementType       = FERequirements_;
-    using ResultRequirementsType  = ResultRequirements<FERequirementType>;
-    using LocalView               = typename FlatBasis::LocalView;
-    using Element                 = typename LocalView::Element;
-    using Traits                  = TraitsFromLocalView<LocalView, useEigenRef>;
-    using Geometry                = typename Element::Geometry;
-    using GridView                = typename FlatBasis::GridView;
-    static constexpr int myDim    = Traits::mydim;
-    static constexpr int worldDim = Traits::worlddim;
-    using LocalBasisType          = decltype(std::declval<LocalView>().tree().child(0).finiteElement().localBasis());
+    using Traits                 = TraitsFromFE<Basis_, FERequirements_, useEigenRef>;
+    using Basis                  = typename Traits::Basis;
+    using FlatBasis              = typename Traits::FlatBasis;
+    using FERequirementType      = typename Traits::FERequirementType;
+    using LocalView              = typename Traits::LocalView;
+    using Geometry               = typename Traits::Geometry;
+    using GridView               = typename Traits::GridView;
+    using Element                = typename Traits::Element;
+    using ResultRequirementsType = typename Traits::ResultRequirementsType;
+    using BaseDisp               = PowerBasisFE<FlatBasis>;  // Handles globalIndices function
+    using VolumeType             = Volume<LinearElastic<Basis_, FERequirements_, useEigenRef>, Traits>;
+    using TractionType           = Traction<LinearElastic<Basis_, FERequirements_, useEigenRef>, Traits>;
+    static constexpr int myDim   = Traits::mydim;
+    using LocalBasisType         = decltype(std::declval<LocalView>().tree().child(0).finiteElement().localBasis());
 
     /**
      * @brief Constructor for the LinearElastic class.
@@ -74,30 +80,27 @@ namespace Ikarus {
     LinearElastic(const Basis& globalBasis, const typename LocalView::Element& element, double emod, double nu,
                   VolumeLoad p_volumeLoad = {}, const BoundaryPatch<GridView>* p_neumannBoundary = nullptr,
                   NeumannBoundaryLoad p_neumannBoundaryLoad = {})
-        : BaseDisp(globalBasis.flat(), element), neumannBoundary{p_neumannBoundary}, emod_{emod}, nu_{nu} {
+        : BaseDisp(globalBasis.flat(), element),
+          VolumeType(p_volumeLoad),
+          TractionType(p_neumannBoundary, p_neumannBoundaryLoad),
+          emod_{emod},
+          nu_{nu} {
       this->localView().bind(element);
       auto& first_child = this->localView().tree().child(0);
       const auto& fe    = first_child.finiteElement();
       geo_              = std::make_shared<const Geometry>(this->localView().element().geometry());
-      numberOfNodes     = fe.size();
-      order             = 2 * (fe.localBasis().order());
+      numberOfNodes_    = fe.size();
+      order_            = 2 * (fe.localBasis().order());
       localBasis        = Dune::CachedLocalBasis(fe.localBasis());
-      if constexpr (requires { this->localView().element().impl().getQuadratureRule(order); })
+      if constexpr (requires { this->localView().element().impl().getQuadratureRule(order_); })
         if (this->localView().element().impl().isTrimmed())
-          localBasis.bind(this->localView().element().impl().getQuadratureRule(order), Dune::bindDerivatives(0, 1));
+          localBasis.bind(this->localView().element().impl().getQuadratureRule(order_), Dune::bindDerivatives(0, 1));
         else
-          localBasis.bind(Dune::QuadratureRules<double, myDim>::rule(this->localView().element().type(), order),
+          localBasis.bind(Dune::QuadratureRules<double, myDim>::rule(this->localView().element().type(), order_),
                           Dune::bindDerivatives(0, 1));
       else
-        localBasis.bind(Dune::QuadratureRules<double, myDim>::rule(this->localView().element().type(), order),
+        localBasis.bind(Dune::QuadratureRules<double, myDim>::rule(this->localView().element().type(), order_),
                         Dune::bindDerivatives(0, 1));
-
-      if constexpr (!std::is_same_v<VolumeLoad, utils::LoadDefault>) volumeLoad = p_volumeLoad;
-      if constexpr (!std::is_same_v<NeumannBoundaryLoad, utils::LoadDefault>)
-        neumannBoundaryLoad = p_neumannBoundaryLoad;
-
-      assert(((not p_neumannBoundary and not neumannBoundaryLoad) or (p_neumannBoundary and neumannBoundaryLoad))
-             && "If you pass a Neumann boundary you should also pass the function for the Neumann load!");
     }
     /**
      * @brief Gets the displacement function for the given FERequirementType and optional displacement vector.
@@ -111,7 +114,7 @@ namespace Ikarus {
     auto displacementFunction(const FERequirementType& par,
                               const std::optional<const Eigen::VectorX<ScalarType>>& dx = std::nullopt) const {
       const auto& d = par.getGlobalSolution(Ikarus::FESolutions::displacement);
-      auto disp     = Ikarus::FEHelper::localSolutionBlockVector<FERequirementType>(d, this->localView(), dx);
+      auto disp     = Ikarus::FEHelper::localSolutionBlockVector<Traits>(d, this->localView(), dx);
       Dune::StandardLocalFunction uFunction(localBasis, disp, geo_);
       return uFunction;
     }
@@ -151,6 +154,10 @@ namespace Ikarus {
       return [&]([[maybe_unused]] auto gp) { return materialTangent(); };
     }
 
+    const Geometry& geometry() const { return *geo_; }
+    [[nodiscard]] size_t numberOfNodes() const { return numberOfNodes_; }
+    [[nodiscard]] int order() const { return order_; }
+
     /**
      * @brief Calculates the scalar energy for the given FERequirementType.
      *
@@ -183,14 +190,18 @@ namespace Ikarus {
       const auto C = materialTangent();
       for (const auto& [gpIndex, gp] : eps.viewOverIntegrationPoints()) {
         const double intElement = geo_->integrationElement(gp.position()) * gp.weight();
-        for (size_t i = 0; i < numberOfNodes; ++i) {
+        for (size_t i = 0; i < numberOfNodes_; ++i) {
           const auto bopI = eps.evaluateDerivative(gpIndex, wrt(coeff(i)), on(gridElement));
-          for (size_t j = 0; j < numberOfNodes; ++j) {
+          for (size_t j = 0; j < numberOfNodes_; ++j) {
             const auto bopJ = eps.evaluateDerivative(gpIndex, wrt(coeff(j)), on(gridElement));
             K.template block<myDim, myDim>(i * myDim, j * myDim) += bopI.transpose() * C * bopJ * intElement;
           }
         }
       }
+
+      // Update due to displacement-dependent external loads, e.g., follower loads
+      VolumeType::calculateMatrix(par, K);
+      TractionType::calculateMatrix(par, K);
     }
 
     /**
@@ -218,17 +229,13 @@ namespace Ikarus {
         DUNE_THROW(Dune::NotImplemented, "The requested result type is NOT implemented.");
     }
 
+  private:
     std::shared_ptr<const Geometry> geo_;
     Dune::CachedLocalBasis<std::remove_cvref_t<LocalBasisType>> localBasis;
-    std::function<Eigen::Vector<double, worldDim>(const Dune::FieldVector<double, worldDim>&, const double&)>
-        volumeLoad;
-    std::function<Eigen::Vector<double, worldDim>(const Dune::FieldVector<double, worldDim>&, const double&)>
-        neumannBoundaryLoad;
-    const BoundaryPatch<GridView>* neumannBoundary;
     double emod_;
     double nu_;
-    size_t numberOfNodes{0};
-    int order{};
+    size_t numberOfNodes_{0};
+    int order_{};
 
   protected:
     template <typename ScalarType>
@@ -245,52 +252,21 @@ namespace Ikarus {
       ScalarType energy = 0.0;
       for (const auto& [gpIndex, gp] : eps.viewOverIntegrationPoints()) {
         const auto EVoigt = eps.evaluate(gpIndex, on(gridElement));
-
         energy += (0.5 * EVoigt.dot(C * EVoigt)) * geo_->integrationElement(gp.position()) * gp.weight();
       }
 
       // External forces volume forces over the domain
-      if (volumeLoad) {
-        for (const auto& [gpIndex, gp] : eps.viewOverIntegrationPoints()) {
-          const auto uVal                              = uFunction.evaluate(gpIndex);
-          Eigen::Vector<double, Traits::worlddim> fext = volumeLoad(geo_->global(gp.position()), lambda);
-          energy -= uVal.dot(fext) * geo_->integrationElement(gp.position()) * gp.weight();
-        }
-      }
+      energy += VolumeType::calculateScalarImpl(par, dx);
 
       // line or surface loads, i.e., neumann boundary
-      if (not neumannBoundary and not neumannBoundaryLoad) return energy;
-
-      auto element = this->localView().element();
-      for (auto&& intersection : intersections(neumannBoundary->gridView(), element)) {
-        if (not neumannBoundary->contains(intersection)) continue;
-
-        const auto& quadLine = Dune::QuadratureRules<double, myDim - 1>::rule(intersection.type(), order);
-
-        for (const auto& curQuad : quadLine) {
-          // Local position of the quadrature point
-          const Dune::FieldVector<double, myDim>& quadPos = intersection.geometryInInside().global(curQuad.position());
-
-          const double integrationElement = intersection.geometry().integrationElement(curQuad.position());
-
-          // The value of the local function
-          const auto uVal = uFunction.evaluate(quadPos);
-
-          // Value of the Neumann data at the current position
-          auto neumannValue = neumannBoundaryLoad(intersection.geometry().global(curQuad.position()), lambda);
-
-          energy -= neumannValue.dot(uVal) * curQuad.weight() * integrationElement;
-        }
-      }
+      energy += TractionType::calculateScalarImpl(par, dx);
       return energy;
     }
 
     template <typename ScalarType>
     void calculateVectorImpl(const FERequirementType& par, typename Traits::template VectorType<ScalarType> force,
                              const std::optional<const Eigen::VectorX<ScalarType>>& dx = std::nullopt) const {
-      const auto& lambda   = par.getParameter(Ikarus::FEParameter::loadfactor);
-      const auto uFunction = displacementFunction(par, dx);
-      const auto eps       = strainFunction(par, dx);
+      const auto eps = strainFunction(par, dx);
       using namespace Dune::DerivativeDirections;
       using namespace Dune;
 
@@ -300,48 +276,17 @@ namespace Ikarus {
       for (const auto& [gpIndex, gp] : eps.viewOverIntegrationPoints()) {
         const double intElement = geo_->integrationElement(gp.position()) * gp.weight();
         auto stresses           = (C * eps.evaluate(gpIndex, on(gridElement))).eval();
-        for (size_t i = 0; i < numberOfNodes; ++i) {
+        for (size_t i = 0; i < numberOfNodes_; ++i) {
           const auto bopI = eps.evaluateDerivative(gpIndex, wrt(coeff(i)), on(gridElement));
           force.template segment<myDim>(myDim * i) += bopI.transpose() * stresses * intElement;
         }
       }
 
       // External forces volume forces over the domain
-      if (volumeLoad) {
-        for (const auto& [gpIndex, gp] : uFunction.viewOverIntegrationPoints()) {
-          Eigen::Vector<double, Traits::worlddim> fext = volumeLoad(geo_->global(gp.position()), lambda);
-          for (size_t i = 0; i < numberOfNodes; ++i) {
-            const auto udCi = uFunction.evaluateDerivative(gpIndex, wrt(coeff(i)));
-            force.template segment<myDim>(myDim * i)
-                -= udCi * fext * geo_->integrationElement(gp.position()) * gp.weight();
-          }
-        }
-      }
+      VolumeType::calculateVectorImpl(par, force, dx);
 
       // External forces, boundary forces, i.e., at the Neumann boundary
-      if (not neumannBoundary) return;
-      auto element = this->localView().element();
-      for (auto&& intersection : intersections(neumannBoundary->gridView(), element)) {
-        if (not neumannBoundary->contains(intersection)) continue;
-
-        // Integration rule along the boundary
-        const auto& quadLine = Dune::QuadratureRules<double, myDim - 1>::rule(intersection.type(), order);
-
-        for (const auto& curQuad : quadLine) {
-          const Dune::FieldVector<double, myDim>& quadPos = intersection.geometryInInside().global(curQuad.position());
-
-          const double integrationElement = intersection.geometry().integrationElement(curQuad.position());
-
-          // The value of the local function wrt the i-th coeff
-          for (size_t i = 0; i < numberOfNodes; ++i) {
-            const auto udCi = uFunction.evaluateDerivative(quadPos, wrt(coeff(i)));
-
-            // Value of the Neumann data at the current position
-            auto neumannValue = neumannBoundaryLoad(intersection.geometry().global(curQuad.position()), lambda);
-            force.template segment<myDim>(myDim * i) -= udCi * neumannValue * curQuad.weight() * integrationElement;
-          }
-        }
-      }
+      TractionType::calculateVectorImpl(par, force, dx);
     }
   };
 }  // namespace Ikarus