+++ ikarus/finiteelements/mechanics/loads.hh
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2021-2024 The Ikarus Developers mueller@ibb.uni-stuttgart.de
+// SPDX-License-Identifier: LGPL-3.0-or-later
+
+/**
+ * @file loads.hh
+ * @brief Header file for types of loads in Ikarus finite element mechanics.
+ * @ingroup  mechanics
+ */
+
+#pragma once
+
+#include <ikarus/finiteelements/mechanics/loads/traction.hh>
+#include <ikarus/finiteelements/mechanics/loads/volume.hh>